+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Breaking changes
 
+* @PrimaryKey field value can be null for String, Byte, Short, Integer, and Long types. Older Realms should be migrated, using RealmObjectSchema.setNullable(), or by adding the @Required annotation. (#2515).
 * RealmResults.clear() now throws UnsupportedOperationException. Use RealmResults.deleteAllFromRealm() instead.
 * RealmResults.remove(int) now throws UnsupportedOperationException. Use RealmResults.deleteFromRealm() instead.
 * Removed deprecated methods Realm.getTable() from public API.